+++ cognite/client/data_classes/events.py
@@ -237,23 +237,6 @@ class EventUpdate(CogniteUpdate):
     # GenStop
 
 
-# GenPropertyClass: EventAggregate
-class EventAggregate(dict):
-    """Aggregation group of events
-
-    Args:
-        count (int): Size of the aggregation group
-    """
-
-    def __init__(self, count: int = None, **kwargs):
-        self.count = count
-        self.update(kwargs)
-
-    count = CognitePropertyClassUtil.declare_property("count")
-
-    # GenStop
-
-
 class EventList(CogniteResourceList):
     _RESOURCE = Event
     _UPDATE = EventUpdate